\name{digeR-package}
\Rdversion{1.1}
\alias{digeR-package}
\docType{package}
\title{
digeR Graphical User Interface
}
\description{
digeR is an easy to use GUI tool for 2D gel or 2D DIGE data analysis. \cr 
It provides functions for spots correlation analysis, score plot, classification, feature selection and power anlaysis
}
\details{
\tabular{ll}{
Package: \tab digeR\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2009-07-26\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}
}
\author{
Yue Fan \email{yue.fan@ucd.ie},Thomas Brendan Murphy, R. William G. Watson

Maintainer: Yue Fan \email{yue.fan@ucd.ie}
}

\keyword{ package }
